requirements.txt
flask
python-dotenv
requests
gunicorn
scikit-learn
joblib
numpy
